refactor: テーマ関連の処理を分離
